Refactor stream reader API
