+++ Tests/LiveKitTests/DataStream/ByteStreamReaderTests.swift
@@ -18,30 +18,114 @@
 import XCTest
 
 class ByteStreamReaderTests: XCTestCase {
-        
-    func testInitialization() {
-        let info = ByteStreamInfo(
-            id: UUID().uuidString,
-            mimeType: "application/octet-stream",
-            topic: "someTopic",
-            timestamp: Date(),
-            totalLength: nil,
-            attributes: [:],
-            name: "filename.bin"
-        )
-        let source = StreamReaderSource { _ in }
-        let reader = ByteStreamReader(info: info, source: source)
-        
-        XCTAssertEqual(reader.info, info)
+    
+    private var continuation: StreamReaderSource.Continuation!
+    private var reader: ByteStreamReader!
+    
+    private let testInfo = ByteStreamInfo(
+        id: UUID().uuidString,
+        mimeType: "application/octet-stream",
+        topic: "someTopic",
+        timestamp: Date(),
+        totalLength: nil,
+        attributes: [:],
+        name: "filename.bin"
+    )
+    
+    let testChunks = [
+        Data(repeating: 0xAB, count: 128),
+        Data(repeating: 0xCD, count: 128),
+        Data(repeating: 0xEF, count: 256),
+        Data(repeating: 0x12, count: 32)
+    ]
+    
+    /// All chunks combined.
+    private var testPayload: Data {
+        testChunks.reduce(Data()) { $0 + $1 }
     }
     
-    /*
+    private func sendPayload(closingError: Error? = nil) {
+        for chunk in testChunks {
+            continuation.yield(chunk)
+        }
+        continuation.finish(throwing: closingError)
+    }
+    
+    override func setUp() {
+        super.setUp()
+        let source = StreamReaderSource {
+            self.continuation = $0
+        }
+        reader = ByteStreamReader(info: testInfo, source: source)
+    }
+    
+    func testChunkRead() async throws {
+        let receiveExpectation = expectation(description: "Receive all chunks")
+        let closureExpectation = expectation(description: "Normal closure")
+        
+        Task {
+            var chunkIndex = 0
+            for try await chunk in reader {
+                XCTAssertEqual(chunk, testChunks[chunkIndex])
+                if chunkIndex == testChunks.count - 1 {
+                    receiveExpectation.fulfill()
+                }
+                chunkIndex += 1
+            }
+            closureExpectation.fulfill()
+        }
+        
+        sendPayload()
+        
+        await fulfillment(
+            of: [receiveExpectation, closureExpectation],
+            timeout: 5,
+            enforceOrder: true
+        )
+    }
+    
+    func testChunkReadError() async throws {
+        let throwsExpectation = expectation(description: "Read throws error")
+        let testError = StreamError.abnormalEnd(reason: "test")
+        
+        Task {
+            do {
+                for try await _ in reader {}
+            } catch {
+                XCTAssertEqual(error as? StreamError, testError)
+                throwsExpectation.fulfill()
+            }
+        }
+        sendPayload(closingError: testError)
+        
+        await fulfillment(
+            of: [throwsExpectation],
+            timeout: 5
+        )
+    }
+    
+    func testReadAll() async throws {
+        let readExpectation = expectation(description: "Read full payload")
+        
+        Task {
+            let fullPayload = try await reader.readAll()
+            XCTAssertEqual(fullPayload, testPayload)
+            readExpectation.fulfill()
+        }
+        sendPayload()
+        
+        await fulfillment(
+            of: [readExpectation],
+            timeout: 5
+        )
+    }
+    
     func testReadToFile() async throws {
         let writtenExpectation = expectation(description: "File properly written")
         Task {
             do {
                 let fileURL = try await reader.readToFile()
-                XCTAssertEqual(fileURL.lastPathComponent, reader.info.fileName)
+                XCTAssertEqual(fileURL.lastPathComponent, reader.info.name)
                 
                 let fileContents = try Data(contentsOf: fileURL)
                 XCTAssertEqual(fileContents, testPayload)
@@ -58,7 +142,6 @@ class ByteStreamReaderTests: XCTestCase {
             timeout: 5
         )
     }
-     */
     
     func testResolveFileName() {
         XCTAssertEqual(
@@ -113,3 +196,4 @@ class ByteStreamReaderTests: XCTestCase {
         )
     }
 }
+
